Implemented Controller sessionPackage

+++ SportCenter-Backend/src/main/java/ca/mcgill/ecse321/sportcenter/service/SessionPackageService.java
@@ -2,6 +2,7 @@ package ca.mcgill.ecse321.sportcenter.service;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import java.sql.Date;
+import java.util.List;
 import ca.mcgill.ecse321.sportcenter.model.Course;
 import ca.mcgill.ecse321.sportcenter.model.SessionPackage;
 import ca.mcgill.ecse321.sportcenter.repository.CourseRepository;
@@ -23,15 +24,17 @@ public class SessionPackageService {
     @Autowired
     private CourseRepository courseRepo;
 
-    @Autowired
-    private SportCenterRepository sportCenterRepository;
-
     //--------------------------// Create SessionPackage //--------------------------//
 
     @Transactional
-    public SessionPackage createSessionPackage(int aPriceReduction, int aDuration, Date aDate, Course aCourse){
-
-        //Input validation
+    public SessionPackage createSessionPackage(int aPriceReduction, int aDuration, Date aDate, int courseId){
+        Course aCourse = courseRepo.findCourseById(courseId);
+        if(aPriceReduction > 100 || aPriceReduction<0 || aDuration <0 || aDate == null ){
+            throw new IllegalArgumentException("Invalid input");
+        }
+        if(aCourse == null){
+            throw new IllegalArgumentException("Course not found");
+        }
         SessionPackage sessionPackage = new SessionPackage(aPriceReduction, aDuration, aDate, aCourse);
         return sessionPackageRepo.save(sessionPackage);
     }
@@ -40,7 +43,14 @@ public class SessionPackageService {
 
     @Transactional
     public SessionPackage updateSessionPackageReduction(int id, int aNewPriceReduction){
-        //Input valid
+        
+        SessionPackage sessionPackage = sessionPackageRepo.findSessionPackageById(id);
+        if(sessionPackage == null){
+            throw new IllegalArgumentException("Session package not found");
+        }
+        if(aNewPriceReduction<0 || aNewPriceReduction >100){
+            throw new IllegalArgumentException("The new price reduction must be between 0 and 100");
+        }
         SessionPackage sessionPackageToUpdate = sessionPackageRepo.findSessionPackageById(id);
         sessionPackageToUpdate.setPriceReduction(aNewPriceReduction);
         return sessionPackageRepo.save(sessionPackageToUpdate);
@@ -50,11 +60,35 @@ public class SessionPackageService {
 
     @Transactional
     public boolean deleteSessionPackage(int id){
-        //Input valid
+        
         SessionPackage sessionPackageToDelete = sessionPackageRepo.findSessionPackageById(id);
+        if(sessionPackageToDelete == null){
+            throw new IllegalArgumentException("Session package not found");
+        }
         sessionPackageRepo.delete(sessionPackageToDelete);
         return true;
     }
+
+    //---------------------------// Getters //-------------------------------//
+
+    @Transactional
+    public List<SessionPackage> getSessionPackagesByCourse(int courseId){
+        Course course = courseRepo.findCourseById(courseId);
+        if(course ==null){
+            throw new IllegalArgumentException("Course not found");
+        }
+        return sessionPackageRepo.findSessionPackageByCourse(course);
+    }
+
+    @Transactional
+    public SessionPackage getSessionPackageById(int Id){
+        SessionPackage sessionPackage = sessionPackageRepo.findSessionPackageById(Id);
+        if(sessionPackage ==null){
+            throw new IllegalArgumentException("SessionPackage not found");
+        }
+        return sessionPackage;
+    }
+
 
 
 }
